+++ src/main/java/org/nmfw/foodietree/domain/CommonController.java
@@ -11,7 +11,10 @@ import org.nmfw.foodietree.domain.customer.dto.resp.CustomerMyPageDto;
 import org.nmfw.foodietree.domain.customer.mapper.CustomerMapper;
 import org.nmfw.foodietree.domain.customer.service.CustomerMyPageService;
 import org.nmfw.foodietree.domain.customer.util.LoginUtil;
+import org.nmfw.foodietree.domain.store.dto.resp.StoreListByEndTimeDto;
+import org.nmfw.foodietree.domain.store.dto.resp.StoreListCo2Dto;
 import org.nmfw.foodietree.domain.store.mapper.StoreMapper;
+import org.nmfw.foodietree.domain.store.service.StoreList.StoreListService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -27,18 +30,35 @@ import org.springframework.web.bind.annotation.PostMapping;
 @RequiredArgsConstructor
 public class CommonController {
 
-    private final CustomerMapper customerMapper;
-    private final StoreMapper storeMapper;
+//    private final CustomerMapper customerMapper;
+//    private final StoreMapper storeMapper;
+//
+//    @GetMapping
+//    public String root(HttpSession session) {
+//        String loggedInUser = LoginUtil.getLoggedInUser(session);
+//        if (loggedInUser != null && customerMapper.findOne(loggedInUser) != null) {
+//            return "redirect:/product/main";
+//        }
+//        if (loggedInUser != null && storeMapper.findOne(loggedInUser) != null) {
+//            return "redirect:/store/mypage/main";
+//        }
+//        return "index";
+//    }
 
+    private final StoreListService storeListService;
+
+    // guest 화면 가게 정보 렌더링
     @GetMapping
-    public String root(HttpSession session) {
-        String loggedInUser = LoginUtil.getLoggedInUser(session);
-        if (loggedInUser != null && customerMapper.findOne(loggedInUser) != null) {
-            return "redirect:/product/main";
-        }
-        if (loggedInUser != null && storeMapper.findOne(loggedInUser) != null) {
-            return "redirect:/store/mypage/main";
-        }
-        return "index";
+    public String getStoreListsForGuests(Model model) {
+        List<StoreListCo2Dto> storesByProductCount = storeListService.getStoresByProductCnt();
+        List<StoreListByEndTimeDto> storesByEndTime = storeListService.getStoresByProductEndTime();
+        // co2를 가장 많이 줄인 순
+        model.addAttribute("storesByProductCount", storesByProductCount);
+        log.info("상품을 가장 많이 판 순 : {}", storesByProductCount);
+        // 상품 시간이 현재로부터 제일 가까운 순 (마감임박)
+        model.addAttribute("storesByEndTime", storesByEndTime);
+        log.info("상품 마감 임박 순 : {}", storesByEndTime);
+
+        return "index"; // JSP 파일명
     }
 }
